pyproject.toml
[project]
name = "ats-automation"
version = "1.0.0"
description = "ATS Automation System - RPA for BPO Clients"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "playwright>=1.50.0",
    "selenium>=4.20.0",
    "pyautogui>=0.9.54",
    "pdfplumber>=0.11.0",
    "pypdf>=5.0.0",
    "pytesseract>=0.3.10",
    "Pillow>=11.0.0",
    "transformers>=4.40.0",
    "torch>=2.3.0",
    "spacy>=3.8.0",
    "numpy>=1.26.0",
    "watchdog>=4.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "loguru>=0.7.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
